old-tests/submission/Microsoft/video/video_026.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'volumechange' event fires when the 'muted' attribute changes.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-volume" />
  <meta name="assert" content="'volumechange' event fires when the 'muted' attribute changes." />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>  
</head>
<body>
  <div id=log></div>
  <video id="video0">Your browser does not support media elements.</video>
  <script type="text/javascript">
  
   
  var test = async_test("'volumechange' event fires when the 'muted' attribute changes.")
  var vid = document.getElementById("video0");
  
  vid.volume = 1.0;
  vid.addEventListener(
      "volumechange",
      function(event){
              test.step(function(){assert_equals(event.type, "volumechange")});
              test.done();
          },
      false
      );
  
  // change muted to a new value
  vid.muted = true;

  function timeout()
  {
    test.step(function(){assert_true(false, "'volumechange' event never fired!")});
    test.done();
  }
  
  var timer = setTimeout(timeout, 4000);
  
  </script>
</body>
</html>
